Add function that searches PR files for list of non-inclusive words
In the future, we'll want to extract out that list to a separate file. It'd be a little more complicated to to check that the actual _word_ is actually being added in the PR (it's not just on a line that's been modified and thus contained in an "added" line), but we should look into that in the future.